unify activation and usability function
