+++ src/layer/arm/deconvolution_arm.cpp
@@ -18,13 +18,9 @@
 
 #if __ARM_NEON
 #include <arm_neon.h>
-#include "neon_mathfun.h"
-#if __ARM_FEATURE_FP16_VECTOR_ARITHMETIC
-#include "neon_mathfun_fp16s.h"
-#endif
 #endif // __ARM_NEON
 
-#include "neon_activation.h"
+#include "arm_activation.h"
 
 namespace ncnn {
