+++ src/layer/x86/prelu_x86.cpp
@@ -14,12 +14,7 @@
 
 #include "prelu_x86.h"
 
-#if __SSE2__
-#include "sse_activation.h"
-#if __AVX__
-#include "avx_activation.h"
-#endif // __AVX__
-#endif // __SSE2__
+#include "x86_activation.h"
 
 namespace ncnn {
