--- src/layer/x86/sse_usability.h
@@ -1,27 +1,0 @@
-// Tencent is pleased to support the open source community by making ncnn available.
-//
-// Copyright (C) 2020 THL A29 Limited, a Tencent company. All rights reserved.
-//
-// Licensed under the BSD 3-Clause License (the "License"); you may not use this file except
-// in compliance with the License. You may obtain a copy of the License at
-//
-// https://opensource.org/licenses/BSD-3-Clause
-//
-// Unless required by applicable law or agreed to in writing, software distributed
-// under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
-// CONDITIONS OF ANY KIND, either express or implied. See the License for the
-// specific language governing permissions and limitations under the License.
-
-#ifndef SSE_USABILITY_H
-#define SSE_USABILITY_H
-
-#include <emmintrin.h>
-
-static inline float _mm_reduce_add_ps(__m128 x128)
-{
-    const __m128 x64 = _mm_add_ps(x128, _mm_movehl_ps(x128, x128));
-    const __m128 x32 = _mm_add_ss(x64, _mm_shuffle_ps(x64, x64, 0x55));
-    return _mm_cvtss_f32(x32);
-}
-
-#endif // SSE_USABILITY_H
